backend/requirements.txt
Django>=5.1,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
psycopg2-binary>=2.9
Pillow>=10.4
requests>=2.32
